Support sending a mail to multiple recipients
This modifies how we interact with Notify to allow a user to send a
single email to multiple recipients, as they would be able to do with
any other mailer.

Notify doesn't actually support sending a single email to multiple
recipients in the way that SMTP supports it, so instead we make a single
call to Notify for each recipient. This slightly changes how we return
the response from Notify to be an array rather than a single object. For
backwards compatibility, I've left the existing `govuk_notify_response`
value as the first response.

The behaviour at the moment is misleading as it simply takes the first
recipient from the list and discards the rest without any warning.